pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lattice-lock"
version = "2.1.0"
description = "A governance-first framework for AI-assisted software development with intelligent model orchestration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lattice Lock Team"}
]
keywords = [
    "ai",
    "llm",
    "orchestration",
    "model-routing",
    "governance",
    "code-generation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "Jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "defusedxml>=0.7.1",
    "aiofiles>=23.0.0",
    "bcrypt>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "boto3>=1.28.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "pylint>=3.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "types-aiofiles>=23.0.0",
    "pip-tools>=7.0.0",
]
all = [
    "uvloop>=0.19.0; platform_system != 'Windows'",
]

[project.urls]
Homepage = "https://github.com/klappe-pm/lattice-lock-framework"
Repository = "https://github.com/klappe-pm/lattice-lock-framework"

[project.scripts]
lattice-lock = "lattice_lock.cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
norecursedirs = [
    "scripts/tests",
    "examples",
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.pytest_cache
    | build
    | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = [".git", ".venv", "__pycache__", ".pytest_cache", "build", "dist"]
known_first_party = ["lattice_lock"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (common in frameworks)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002", "ARG003", "S101", "F841", "B017", "E402", "E712", "SIM103", "SIM117"]  # Allow assert, unused args/vars, and style issues in tests
"scripts/*" = ["ARG001", "ARG002", "F401", "F841", "E402", "B007", "B017", "SIM103", "SIM105", "SIM117", "E712"]  # Allow unused imports/vars and style issues in scripts
"examples/*" = ["F401", "F841", "ARG002", "E402", "B007", "B904", "SIM102", "SIM105", "SIM117"]  # Allow style issues in examples
"src/*" = ["ARG002", "ARG003", "SIM102", "SIM103", "SIM105", "SIM108", "SIM117", "SIM118", "B007", "B904", "C401", "E402"]  # Allow unused method args and style preferences in src

# Pylint configuration
[tool.pylint.main]
ignore = [".git", ".venv", "__pycache__", ".pytest_cache", "build", "dist"]
ignore-patterns = ["test_.*\\.py"]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring (we'll add these in A4)
    "C0115",  # missing-class-docstring (we'll add these in A4)
    "C0116",  # missing-function-docstring (we'll add these in A4)
    "R0903",  # too-few-public-methods
    "W1203",  # logging-fstring-interpolation
]

[tool.pylint.format]
max-line-length = 100
